Move syscall tests to their own file.
